+++ PAWpy/Services/__init__.py
@@ -6,6 +6,11 @@ from PAWpy.Services.ViewService import ViewService
 from PAWpy.Services.AdminService import AdminService
 from PAWpy.Services.UIService import UIService
 from PAWpy.Services.TM1ProxyService import TM1ProxyService
+from PAWpy.Services.ContentV1Service import ContentV1Service
+from PAWpy.Services.UserGroupService import UserGroupService
+from PAWpy.Services.DatabaseService import DatabaseService
+from PAWpy.Services.UserAdminService import UserAdminService
+from PAWpy.Services.CloudAdminService import CloudAdminService
 
 __all__ = [
     "RestService",
@@ -16,4 +21,9 @@ __all__ = [
     "AdminService",
     "UIService",
     "TM1ProxyService",
+    "ContentV1Service",
+    "UserGroupService",
+    "DatabaseService",
+    "UserAdminService",
+    "CloudAdminService",
 ]